move qnode_call to the workflow module and make it private
